правка миграции

+++ src/services/SitemapService.php
@@ -154,6 +154,9 @@ class SitemapService extends Service
             'entity_type_id' => $entity_type->id,
             'entity_id' => $model->id,
         ]);
+        if (! $current){
+            return true;
+        }
 
         return $this->delete($current);
     }
